+++ .idea/compiler.xml
@@ -7,6 +7,7 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
+        <module name="CalendaringMicroservice" />
         <module name="Blackbird" />
         <module name="FreeWoman" />
       </profile>
@@ -15,6 +16,7 @@
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
       <module name="Blackbird" options="-parameters" />
+      <module name="CalendaringMicroservice" options="-parameters" />
       <module name="FreeWoman" options="-parameters" />
     </option>
   </component>
