+++ src/main/java/com/hfla/service/calendar/controllers/CalendarController.java
@@ -4,12 +4,13 @@ import java.io.IOException;
 import java.time.Instant;
 import java.util.List;
 
+import com.hfla.service.calendar.pojos.CalendarsInterface;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.hfla.service.calendar.services.CalendarService;
+import com.hfla.service.calendar.services.NylasCalendarService;
 import com.hfla.service.calendar.services.EventService;
 import com.nylas.Event;
 import com.nylas.FreeBusy;
@@ -24,33 +25,33 @@ import com.nylas.Calendar;
 @RequestMapping(path = "/calendars")
 public class CalendarController {
 
-  private final CalendarService calendarService;
+  private final NylasCalendarService nylasCalendarService;
   private final EventService eventService;
 
   @Autowired
-  public CalendarController(CalendarService calendarService, EventService eventService) {
-    this.calendarService = calendarService;
+  public CalendarController(NylasCalendarService nylasCalendarService, EventService eventService) {
+    this.nylasCalendarService = nylasCalendarService;
     this.eventService = eventService;
   }
 
   @GetMapping
-  public RemoteCollection<Calendar> getCalendars() throws IOException, RequestFailedException {
+  public CalendarsInterface getCalendars() throws IOException, RequestFailedException {
     System.out.println("Getting calendars");
-    return calendarService.getCalendars();
+    return nylasCalendarService.getCalendars();
   }
 
   @GetMapping(path = "/freebusy")
   public List<FreeBusy> getFreeBusy() throws IOException, RequestFailedException {
     System.out.println("Getting the free busy");
 
-    return calendarService.checkFreeBusy();
+    return nylasCalendarService.checkFreeBusy();
   }
 
   // TODO: getAvailability should require start date and enddate.
   @GetMapping(path = "/availability")
   public List<TimeSlot> getAvailability() throws IOException, RequestFailedException {
     System.out.println("Getting availability");
-    return calendarService.checkAvailability(Instant.parse("0"), Instant.parse("1"));
+    return nylasCalendarService.checkAvailability(Instant.parse("0"), Instant.parse("1"));
   }
 
   @GetMapping(path = "/events")
