+++ src/main/java/com/hfla/service/calendar/legacy/controllers/ConnectedCalendarController.java
@@ -1,14 +1,11 @@
 package com.hfla.service.calendar.legacy.controllers;
 
-import org.joda.time.DateTime;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.hfla.service.calendar.legacy.pojos.Availability;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import com.hfla.service.calendar.legacy.pojos.Availability;
 import com.hfla.service.calendar.legacy.pojos.ConnectedCalendar;
 import com.hfla.service.calendar.legacy.pojos.ConnectedEvent;
 
@@ -16,11 +13,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.LogManager;
 import java.util.logging.Logger;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 @RestController
-@RequestMapping(path = "/calendar")
+@RequestMapping(path = "/com")
 public class ConnectedCalendarController {
 	//@Autowired
 	Availability availability;
