+++ src/main/java/com/hfla/service/calendar/legacy/pojos/ConnectedCalendar.java
@@ -1,8 +1,8 @@
 package com.hfla.service.calendar.legacy.pojos;
 
+import com.hfla.service.calendar.legacy.clients.NylasRequest;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.hfla.service.calendar.legacy.clients.NylasRequest;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
