+++ src/main/java/com/hfla/service/calendar/legacy/pojos/MeetingType.java
@@ -1,6 +1,5 @@
 package com.hfla.service.calendar.legacy.pojos;
 
-import lombok.Generated;
 import lombok.Getter;
 import lombok.Setter;
