+++ src/main/java/com/hfla/service/calendar/legacy/repositories/MeetingTypeRepository.java
@@ -1,11 +1,8 @@
 package com.hfla.service.calendar.legacy.repositories;
 
+import com.hfla.service.calendar.legacy.pojos.MeetingType;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-
-import com.hfla.service.calendar.legacy.pojos.MeetingType;
-
-import java.util.List;
 
 @Repository
 public interface MeetingTypeRepository extends CrudRepository<MeetingType, Long> {
