+++ src/main/java/com/hfla/service/calendar/pojos/CalendarInterface.java
@@ -1,0 +1,4 @@
+package com.hfla.service.calendar.pojos;
+
+public interface CalendarInterface {
+}
