+++ src/main/java/com/hfla/service/calendar/pojos/Nylas/NylasCalendar.java
@@ -1,0 +1,7 @@
+package com.hfla.service.calendar.pojos.Nylas;
+
+import com.hfla.service.calendar.pojos.CalendarInterface;
+import com.nylas.Calendar;
+
+public class NylasCalendar extends Calendar implements CalendarInterface {
+}
