+++ src/main/java/com/hfla/service/calendar/services/CronofyService.java
@@ -1,18 +1,19 @@
 package com.hfla.service.calendar.services;
 
+import com.hfla.service.calendar.pojos.Cronify.Calendars;
+import com.hfla.service.calendar.pojos.Cronify.Event;
+import com.hfla.service.calendar.pojos.EventsInteface;
+import com.hfla.service.calendar.pojos.Cronify.Events;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import com.hfla.service.calendar.pojos.Calendars;
-import com.hfla.service.calendar.pojos.Event;
-import com.hfla.service.calendar.pojos.Events;
 import reactor.core.publisher.Mono;
 
 @Service
-public class CronofyService {
+public class CronofyService implements CalendarService{
 
   private final WebClient webClient;
 
@@ -24,7 +25,7 @@ public class CronofyService {
         .defaultHeader(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken).build();
   }
 
-  public Events getEvents() {
+  public EventsInteface getEvents() {
     return this.webClient.get().uri("/v1/events?tzid=Etc/UTC").retrieve().bodyToMono(Events.class)
         .block();
   }
@@ -33,7 +34,12 @@ public class CronofyService {
     return this.webClient.get().uri("/v1/calendars").retrieve().bodyToMono(Calendars.class).block();
   }
 
-  public String createEvent(Event event) {
+
+
+
+  @Override
+  public String createEvent(EventsInteface eventParameter) {
+    Events event = (Events)eventParameter;
     String calendarId = this.getCalendars().getCalendars().get(0).getCalendarId();
 
     return this.webClient.post()
