+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=8080
+server.port=8090
 
 ## default connection pool
 spring.datasource.hikari.connectionTimeout=20000
@@ -39,3 +39,4 @@ google.client.id=${GOOGLE_CLIENT_ID}
 
 # Cronofy credentials
 cronofy.access.token=${CRONOFY_ACCESS_TOKEN}
+
